+++ cypress-shared/integration/b2b/2.4-add-remaining-users.spec.js
@@ -16,7 +16,7 @@ import { loginToStoreFront } from '../../support/b2b/login.js'
 
 const { users, gmailCreds } = b2b.OrganizationA
 
-describe('Add binding which shows Organization in profile page, Sync Checkout UI Custom & Add Sales Users via Graphql', () => {
+describe('Sync Checkout UI Custom & Add Sales Users via Graphql', () => {
   loginViaCookies({ storeFrontCookie: false })
 
   loginToStoreFront(